+++ src/main/java/racingcar/controller/WebController.java
@@ -4,8 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import racingcar.dto.PlayRequest;
-import racingcar.dto.PlayResponse;
+import racingcar.request.PlayRequest;
+import racingcar.response.PlayResponse;
 import racingcar.service.GameService;
 
 @RestController
