+++ src/main/java/racingcar/domain/MovableStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public interface MovableStrategy {
