+++ src/main/java/racingcar/domain/TrialCount.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public class TrialCount {
