+++ src/main/java/racingcar/service/GameService.java
@@ -6,12 +6,12 @@ import static java.util.stream.Collectors.toUnmodifiableList;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import racingcar.dto.PlayResponse;
+import racingcar.response.PlayResponse;
 import racingcar.dto.VehicleDto;
-import racingcar.model.Car;
-import racingcar.model.Cars;
-import racingcar.model.TrialCount;
-import racingcar.model.Vehicle;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
+import racingcar.domain.TrialCount;
+import racingcar.domain.Vehicle;
 import racingcar.repository.GameDao;
 import racingcar.util.CarNameValidator;
 import java.util.Arrays;
