+++ src/test/java/racingcar/domain/TrialCountTest.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.assertThatNoException;
